Rebuilt rgbeffects file
I would suggest using this instead of the default rgbeffects file because when working in the sequencer there is *much* less lag.  Also, there are quite a few groups that have been created which allow more effects such as the single line.  This better aligns with how xLights was traditionally used on the house model.

The only sacrifice in this layout is that it breaks the sample show.  If needed I can fix that too but a quick import will address that.